divided suggestion into several components
