+++ package.json
@@ -13,6 +13,7 @@
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "@types/react-redux": "^7.1.33",
+    "@types/react-virtualized": "^9.21.30",
     "clsx": "^2.1.1",
     "firebase": "^10.11.1",
     "react": "^18.3.1",
@@ -24,6 +25,7 @@
     "react-scripts": "5.0.1",
     "react-spinners": "^0.13.8",
     "react-toastify": "^10.0.5",
+    "react-virtualized": "^9.22.5",
     "sass": "^1.75.0",
     "typescript": "^4.4.2",
     "use-debounce": "^10.0.0",
