+++ src/api/redux/api/card-api.ts
@@ -42,7 +42,11 @@ export const cardApi = createApi({
     getSeveralCharectersById: builder.query<CharecterCard[], string>({
       query: (ids) => `/character/${ids}`,
       transformResponse: (response: Result[]) => {
+        console.log(response);
         return parseLoadedCharecters(response);
+      },
+      transformErrorResponse(baseQueryReturnValue, meta, arg) {
+        return [];
       },
     }),
   }),
