+++ src/shared/components/UI/Search/Search.tsx
@@ -5,66 +5,11 @@ import { IconSearch } from "../../../icons/SearchIcon";
 import useSearch from "../../../hooks/useSearch";
 import { useFindCharactersByNameQuery } from "../../../../api/redux/api/card-api";
 import { useDebounce } from "use-debounce";
-import { ClipLoader } from "react-spinners";
-import { NothingYet } from "../../NothingYet/NothingYet";
-import { CharecterCard } from "../../Card/Card";
+import { SuggestList } from "../../SuggestList/SuggestList";
 
 const isSearchBarVisible = (pathname: string): boolean => {
   if (pathname === "/signin" || pathname === "/signup") return false;
   return true;
-};
-
-interface SuggestItemProps {
-  image: string;
-  name: string;
-}
-
-const SuggestItem = ({ image, name }: SuggestItemProps) => {
-  return (
-    <div className={styles.itemWrapper}>
-      <div className={styles.img}>
-        <img src={image} alt="character" />
-      </div>
-      <span>{name}</span>
-    </div>
-  );
-};
-
-interface SuggestListProps {
-  data: CharecterCard[] | undefined;
-  value: string | null;
-  isFetching: boolean;
-  isSucces: boolean;
-  isLoading?: boolean;
-  isError?: boolean;
-}
-
-const SuggestList = ({
-  data,
-  value,
-  isFetching,
-  isSucces,
-  isError,
-  isLoading,
-}: SuggestListProps) => {
-  return (
-    <>
-      <div className={styles.suggestList}>
-        {isFetching || !value ? (
-          <div className={styles.loader}>
-            <ClipLoader color="#36d7b7" />
-          </div>
-        ) : !isSucces ? (
-          <div className={styles.notFound}>
-            <span>Not found</span>
-          </div>
-        ) : null}
-        {!isError && data
-          ? data.map((el, i) => <SuggestItem key={i} {...el} />)
-          : null}
-      </div>
-    </>
-  );
 };
 
 export const Search = () => {
@@ -75,11 +20,9 @@ export const Search = () => {
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
+    setIsVisible(false);
     toggleSearch();
   };
-
-  const { data, isFetching, isSuccess, isLoading, isError } =
-    useFindCharactersByNameQuery(value ?? "", { skip: value ? false : true });
 
   return (
     <>
@@ -87,7 +30,7 @@ export const Search = () => {
         <form onSubmit={handleSubmit} className={styles.searchForm} action="">
           <input
             onFocus={() => setIsVisible(true)}
-            onBlur={() => setIsVisible(false)}
+            // onBlur={() => setIsVisible(false)}
             type="text"
             placeholder="Start typing"
             className={styles.searchInpt}
@@ -97,16 +40,7 @@ export const Search = () => {
           <button type="submit" className={styles.findBtn}>
             <IconSearch />
           </button>
-          {searchQuery && isVisible && (
-            <SuggestList
-              data={data}
-              value={value}
-              isFetching={isFetching}
-              isSucces={isSuccess}
-              isLoading={isLoading}
-              isError={isError}
-            />
-          )}
+          {searchQuery && isVisible && <SuggestList value={value} />}
         </form>
       ) : null}
     </>
